Move presentation logic to PresentInWindowSuggestionPresenter
